Move internal packages to internal directory
