+++ internal/chart/chart_cmd.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/form3tech-oss/f1/internal/support/errorh"
 
-	"github.com/form3tech-oss/f1/pkg/f1/trigger/api"
+	"github.com/form3tech-oss/f1/internal/trigger/api"
 
 	"github.com/wcharczuk/go-chart"
