+++ internal/fluentd_hook/hook.go
