+++ internal/run/test_runner.go
@@ -12,23 +12,23 @@ import (
 	"text/template"
 	"time"
 
-	"github.com/form3tech-oss/f1/pkg/f1/logging"
-	"github.com/form3tech-oss/f1/pkg/f1/options"
+	"github.com/form3tech-oss/f1/internal/logging"
+	"github.com/form3tech-oss/f1/internal/options"
 	"github.com/form3tech-oss/f1/pkg/f1/scenarios"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 
-	"github.com/form3tech-oss/f1/pkg/f1/trace"
+	"github.com/form3tech-oss/f1/internal/trace"
 
-	"github.com/form3tech-oss/f1/pkg/f1/raterun"
+	"github.com/form3tech-oss/f1/internal/raterun"
 
-	"github.com/form3tech-oss/f1/pkg/f1/trigger/api"
+	"github.com/form3tech-oss/f1/internal/trigger/api"
 
 	"github.com/aholic/ggtimer"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/push"
 
-	"github.com/form3tech-oss/f1/pkg/f1/metrics"
+	"github.com/form3tech-oss/f1/internal/metrics"
 	"github.com/form3tech-oss/f1/pkg/f1/testing"
 )
