+++ internal/trigger/configure.go
@@ -1,0 +1,22 @@
+package trigger
+
+import (
+	"github.com/form3tech-oss/f1/internal/trigger/api"
+	"github.com/form3tech-oss/f1/internal/trigger/constant"
+	"github.com/form3tech-oss/f1/internal/trigger/file"
+	"github.com/form3tech-oss/f1/internal/trigger/gaussian"
+	"github.com/form3tech-oss/f1/internal/trigger/ramp"
+	"github.com/form3tech-oss/f1/internal/trigger/staged"
+	"github.com/form3tech-oss/f1/internal/trigger/users"
+)
+
+func GetBuilders() []api.Builder {
+	return []api.Builder{
+		constant.ConstantRate(),
+		staged.StagedRate(),
+		gaussian.GaussianRate(),
+		users.UsersRate(),
+		ramp.RampRate(),
+		file.FileRate(),
+	}
+}
