+++ internal/trigger/constant/constant_rate.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/form3tech-oss/f1/pkg/f1/trigger/api"
+	"github.com/form3tech-oss/f1/internal/trigger/api"
 
 	"github.com/asaskevich/govalidator"
 	"github.com/spf13/pflag"
