+++ internal/trigger/file/file_parser.go
@@ -4,12 +4,12 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/form3tech-oss/f1/pkg/f1/trigger/staged"
+	"github.com/form3tech-oss/f1/internal/trigger/staged"
 
-	"github.com/form3tech-oss/f1/pkg/f1/trigger/ramp"
+	"github.com/form3tech-oss/f1/internal/trigger/ramp"
 
-	"github.com/form3tech-oss/f1/pkg/f1/trigger/constant"
-	"github.com/form3tech-oss/f1/pkg/f1/trigger/gaussian"
+	"github.com/form3tech-oss/f1/internal/trigger/constant"
+	"github.com/form3tech-oss/f1/internal/trigger/gaussian"
 	"gopkg.in/yaml.v2"
 )
