+++ internal/trigger/file/file_parser_test.go
