+++ internal/trigger/staged/stage.go
