+++ pkg/f1/testing/t.go
@@ -7,7 +7,7 @@ import (
 
 	log "github.com/sirupsen/logrus"
 
-	"github.com/form3tech-oss/f1/pkg/f1/metrics"
+	"github.com/form3tech-oss/f1/internal/metrics"
 
 	"github.com/stretchr/testify/require"
 )
